Extracts library type extensions into their own file.
