+++ Borkle/Borkle/BubbleCanvas.swift
@@ -77,10 +77,3 @@ class BubbleCanvas: NSView {
         bezierPath.stroke()
     }
 }
-
-
-extension CGRect {
-    var center: CGPoint {
-        CGPoint(x: midX, y: midY)
-    }
-}
